Add a few explicit conversions back from int
Suppresses GCC's -Wconversion, which is not enabled by default.

error: conversion from ‘int’ to ‘quint8’ {aka ‘unsigned char’} may change value [-Werror=conversion]

Fixes: QTBUG-103001
(cherry picked from commit 214df31916c53e46348173fc02fe53f2b9db729a)

+++ src/corelib/tools/qtaggedpointer.h
@@ -48,7 +48,7 @@
 QT_BEGIN_NAMESPACE
 
 namespace QtPrivate {
-    constexpr quint8 nextByteSize(quint8 bits) { return (bits + 7) / 8; }
+    constexpr quint8 nextByteSize(quint8 bits) { return quint8((bits + 7) / 8); }
 
     template <typename T>
     struct TagInfo
